requirements.txt
altair
numpy
pandas
pydeck
streamlit
matplotlib.pyplot
seaborn
plotly.express